+++ docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Features
+- Added the enclave's creation time info which can be obtained through the `enclave ls` and the `enclave inspect` commands
 
 # 0.52.3
 
@@ -8,7 +10,6 @@
 # 0.52.2
 
 # 0.52.1
-
 ### Features
 - Add `wait` and `define` command in Startosis
 - Added `not found service GUIDs information` in `KurtosisContext.GetServiceLogs` method