+++ docs/changelog.md
@@ -1,6 +1,12 @@
 # TBD
+
 ### Features
 - Added the enclave's creation time info which can be obtained through the `enclave ls` and the `enclave inspect` commands
+- Moved `kurtosis startosis exec` to `kurtosis exec`
+
+### Breaking Features
+- Moved `kurtosis startosis exec` to `kurtosis exec`
+  - Users now need to use the new command to launch Starlark programs
 
 # 0.52.3